services/looker/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.5
Flask-Testing==0.7.1
gunicorn==19.9.0